Moved loadPrices-function to tokenStorage independent class
