+++ src/Resources/config/services.yml
@@ -4,13 +4,19 @@ services:
         factory: 'best_it_ct_customer_prices.model_customer_price_collection.by_user_factory:createPriceCollection'
         lazy: true
 
-    best_it_ct_customer_prices.model_customer_price_collection.by_user_factory:
-        class: BestIt\CtCustomerPricesBundle\Model\CustomerPriceCollection\ByUserFactory
+    best_it_ct_customer_prices.model_customer_price_collection.customer_price_collection_factory:
+        class: BestIt\CtCustomerPricesBundle\Model\CustomerPriceCollection\CustomerPriceCollectionFactory
         arguments:
             - '@best_it_ct_customer_prices.cache_adapter'
             - '%best_it_ct_customer_prices.fields%'
             - '%best_it_ct_customer_prices.query%'
             - '@best_it_ct_customer_prices.commercetools_client'
             - '%best_it_ct_customer_prices.container%'
+        public: false
+
+    best_it_ct_customer_prices.model_customer_price_collection.by_user_factory:
+        class: BestIt\CtCustomerPricesBundle\Model\CustomerPriceCollection\ByUserFactory
+        arguments:
+            - '@best_it_ct_customer_prices.model_customer_price_collection.customer_price_collection_factory'
             - '@security.token_storage'
         public: false
