+++ tests/Model/CustomerPriceCollection/ByUserFactoryTest.php
@@ -2,21 +2,13 @@
 
 namespace Tests\BestIt\CtCustomerPricesBundle\CustomerPriceCollection;
 
-use ArrayObject;
 use BestIt\CtCustomerPricesBundle\Model\CustomerInterface;
 use BestIt\CtCustomerPricesBundle\Model\CustomerPriceCollection;
 use BestIt\CtCustomerPricesBundle\Model\CustomerPriceCollection\ByUserFactory;
-use Commercetools\Commons\Helper\QueryHelper;
-use Commercetools\Core\Client;
-use Commercetools\Core\Model\CustomObject\CustomObject;
-use Commercetools\Core\Request\CustomObjects\CustomObjectQueryRequest;
-use Commercetools\Core\Request\Query\MultiParameter;
+use BestIt\CtCustomerPricesBundle\Model\CustomerPriceCollection\CustomerPriceCollectionFactory;
 use PHPUnit\Framework\TestCase;
-use Psr\Cache\CacheItemInterface;
-use ReflectionObject;
-use Symfony\Component\Cache\Adapter\AdapterInterface;
-use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
 use Symfony\Component\Security\Core\Authentication\Token\TokenInterface;
+use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
 
 /**
  * Test for ByUserFactory.
@@ -61,148 +53,38 @@ class ByUserFactoryTest extends TestCase
     public function testCreatePriceCollectionWithoutUser()
     {
         $fixture = new ByUserFactory(
-            $cacheMock = $this->createMock(AdapterInterface::class),
-            $this->fields,
-            $this->query,
-            $clientMock = $this->createMock(Client::class),
-            $containerName = (string)random_int(1000, 9999),
-            $tokenStorageMock = $this->createMock(TokenStorageInterface::class)
+            $this->createMock(CustomerPriceCollectionFactory::class),
+            $this->createMock(TokenStorageInterface::class)
         );
 
         self::assertInstanceOf(CustomerPriceCollection::class, $fixture->createPriceCollection());
     }
 
     /**
-     * Test to load prices from cache function.
+     * Test for createPriceCollection function.
      */
-    public function testLoadCachedPrices()
+    public function testCreatePriceCollectionWithUser()
     {
         $fixture = new ByUserFactory(
-            $cacheMock = $this->createMock(AdapterInterface::class),
-            $this->fields,
-            $this->query,
-            $clientMock = $this->createMock(Client::class),
-            $containerName = (string)random_int(1000, 9999),
-            $tokenStorageMock = $this->createMock(TokenStorageInterface::class)
+            $customerPriceCollectionFactory = $this->createMock(CustomerPriceCollectionFactory::class),
+            $tokenStorage = $this->createMock(TokenStorageInterface::class)
         );
 
-        $priceCollectionMock = $this->createMock(CustomerPriceCollection::class);
+        $user = $this->createMock(CustomerInterface::class);
 
-        $cacheItemMock = $this->createMock(CacheItemInterface::class);
-        $cacheItemMock
-            ->method('isHit')
-            ->willReturn(true);
-        $cacheItemMock
-            ->method('get')
-            ->willReturn($priceCollectionMock);
-
-        $cacheMock
-            ->method('getItem')
-            ->willReturn($cacheItemMock);
-
-        $userMock = $this->createMock(CustomerInterface::class);
-
-        $tokenMock = $this->createMock(TokenInterface::class);
-        $tokenMock
+        $token = $this->createMock(TokenInterface::class);
+        $token
             ->method('getUser')
-            ->willReturn($userMock);
+            ->willReturn($user);
 
-        $tokenStorageMock
+        $tokenStorage
             ->method('getToken')
-            ->willReturn($tokenMock);
+            ->willReturn($token);
+        $customerPriceCollectionFactory
+            ->method('loadPrices')
+            ->with($user)
+            ->willReturn(new CustomerPriceCollection());
 
-
-        self::assertSame($priceCollectionMock, $fixture->createPriceCollection());
-    }
-
-    /**
-     * Test to load prices.
-     */
-    public function testLoadPrices()
-    {
-        $fixture = new ByUserFactory(
-            $cacheMock = $this->createMock(AdapterInterface::class),
-            $this->fields,
-            $this->query,
-            $clientMock = $this->createMock(Client::class),
-            $containerName = (string)random_int(1000, 9999),
-            $tokenStorageMock = $this->createMock(TokenStorageInterface::class),
-            $queryHelperMock = $this->createMock(QueryHelper::class)
-        );
-
-        $cacheItemMock = $this->createMock(CacheItemInterface::class);
-        $cacheItemMock
-            ->method('isHit')
-            ->willReturn(false);
-        $cacheItemMock
-            ->method('set')
-            ->with(self::callback(
-                function (CustomerPriceCollection $collection) {
-                    $price = $collection->getByArticle('123');
-
-                    self::assertSame('EUR', $price->getValue()->getCurrencyCode());
-                    self::assertSame(1000, $price->getValue()->getCentAmount());
-
-                    return true;
-                }
-            ))
-            ->willReturn($cacheItemMock);
-        $cacheItemMock
-            ->method('expiresAfter')
-            ->with(ByUserFactory::DEFAULT_CACHE_TIME)
-            ->willReturn($cacheItemMock);
-        $cacheItemMock
-            ->method('get')
-            ->willReturn($priceCollection = $this->createMock(CustomerPriceCollection::class));
-
-        $cacheMock
-            ->method('getItem')
-            ->willReturn($cacheItemMock);
-
-        $userMock = $this->createMock(CustomerInterface::class);
-        $userMock
-            ->method('getCustomerIdForArticlePrices')
-            ->willReturn($customerId = (string)random_int(1000, 9999));
-        $userMock
-            ->method('getCustomerCurrencyForArticlePrices')
-            ->willReturn($currency = (string)random_int(1000, 9999));
-
-        $customObject = CustomObject::fromArray([
-            'key' => '123',
-            'value' => [
-                $this->fields['article'] => '123',
-                $this->fields['prices'] => [
-                    'value' => [
-                        'centAmount' => 1000,
-                        'currencyCode' => 'EUR'
-                    ]
-                ],
-            ]
-        ]);
-
-        $queryHelperMock
-            ->method('getAll')
-            ->with(
-                $clientMock,
-                self::callback(
-                    function (CustomObjectQueryRequest $request) use ($containerName, $currency, $customerId) {
-                        $reflectionObject = new ReflectionObject($request);
-                        $paramsProperty = $reflectionObject->getProperty('params');
-                        $paramsProperty->setAccessible(true);
-                        /** @var MultiParameter[] $params */
-                        $params = $paramsProperty->getValue($request);
-
-                        self::assertSame(
-                            'container="' . $containerName . '-'. $currency .'-'. $customerId . '"',
-                            current($params)->getValue()
-                        );
-
-                        return true;
-                    }
-                )
-            )
-            ->willReturn(new ArrayObject([$customObject]));
-
-        self::assertSame($priceCollection, $fixture->loadPrices($userMock));
+        self::assertInstanceOf(CustomerPriceCollection::class, $fixture->createPriceCollection());
     }
 }
